Update MOH_OHSMar28_2023_ratinakere ne eneken shortened sentences.txt
Brian requested that we remove all longer sentences from this dataset.

Some notes about this dataset that he mentioned.    He  compared this to "Shakespeare" old English.  The vocabulary used in this dataset is not really used in the spoken world.  Many of the small word in the text makes it very difficult to speak out loud in a natural way.